+++ v2/CHANGELOG.md
@@ -17,6 +17,7 @@
 - multi_delimiter analyzer support
 - Wildcard analyzer support
 - Backup API support
+- Admin Cluster API support
 
 
 ## [2.0.3](https://github.com/arangodb/go-driver/tree/v2.0.3) (2023-10-31)